Set up views for commenting on answers
This commit allows a user to click the reply button on an answer, see the form, and reply to that answer with a comment. Comments are displayed nested with the answer. This likely causes N+1 queries right now. Commenting on a Question is not possible at this stage but will be reimplemented on the next release